Split structure compilation out to chunk_post_...
